+++ src/LoggerModem.cpp
@@ -9,6 +9,7 @@
  */
 
 #include "LoggerModem.h"
+#include "LoggerBase.h"
 
 // Initialize the static members
 int16_t loggerModem::_priorRSSI           = -9999;
@@ -472,12 +473,12 @@ uint32_t loggerModem::parseNISTBytes(byte nistBytes[4]) {
            '=', String(secFrom1900, BIN));
 
     // Return the timestamp
-    uint32_t unixTimeStamp = secFrom1900 - 2208988800;
+    uint32_t unixTimeStamp = secFrom1900 + epochStart::nist_epoch;
     MS_DBG(F("Unix Timestamp returned by NIST (UTC):"), unixTimeStamp);
     // If before Jan 1, 2019 or after Jan 1, 2030, most likely an error
-    if (unixTimeStamp < 1546300800) {
+    if (unixTimeStamp < EARLIEST_SANE_UNIX_TIMESTAMP) {
         return 0;
-    } else if (unixTimeStamp > 1893456000) {
+    } else if (unixTimeStamp > LATEST_SANE_UNIX_TIMESTAMP) {
         return 0;
     } else {
         return unixTimeStamp;